chore(release): ship v1

+++ MothersonBoxManagement/Data/Dtos/CreateBoxDto.cs
@@ -6,8 +6,8 @@ namespace MothersonBoxManagement.Data.Dtos;
 public class CreateBoxDto
 {
     public BoxType Type { get; set; }
-    public int Height { get; set; }
-    public int Width { get; set; }
-    public int Depth { get; set; }
+    public decimal Height { get; set; }
+    public decimal Width { get; set; }
+    public decimal Depth { get; set; }
     public int ExpectedQuantity { get; set; }
 }
